add 'allow high dpi' flag to SDL create window call
On macOS 11 with FULLSCREEN_DESKTOP mode, without the HIGHDPI flag the
OpenGL viewport gets initialised in the bottom left corner, and there's
apparently no way to get the full DPI-resolved size of the screen in
order to letter/pillarbox things ourselves. SDL_GL_GetDrawableSize()
just returns the dimensions that we passed to SDL_CreateWindow(). With
the HIGHDPI flag though, the GL viewport stretches across the full
screen, but also I can't then explicitly set it with glViewport to
correct the aspect. So I'm boned either way at the moment.